+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 3.0.0
+- Add BundleAnalyzerPlugin plugin
+- Optimize common chunks
+
+## 2.18.3
+- Remove binding regenerator-runtime in global
+- Update dependencies
+
 ## 2.15.0
 - Update react-router
 - Update clientFetchData to manually call the routeOnChange handler